+ palette_test_pdf.py; test palettes for pdfcairo
Complementary to probing Kenneth Moreland's palettes already
converted as rounded .plt with the pngcairo terminal by
gnuplot, the pdfcairo terminal is tested as well.  This will
offer a resolution independent, vector-based access of the
diagnosis.  It is not so much /only/ accessing the results
of the diagnosis as that .pdf output of the plots eventually
plot with gnuplot in the resolution independent .pdf format
will be the targeted format this prepares.