update README[.org,.pdf] past inferno, smooth-cool-warm
Kenneth Moreland revised the two palette types' RGB floating
number definitions.  As result by him, the RGB tupels remain
in the range of [0,1].  This offers to process these palettes
in the same way, as already applied on the other palettes --
which is done.

The README[.org,.pdf] now include the two palettes previously
skipped.  Some edit on the text's content, a little on the
markup used, at the same time.